requirements.txt
fastapi
uvicorn[standard]
SQLAlchemy>=2.0
pydantic>=2.6
python-dotenv
python-multipart
